+++ CHANGELOG.md
@@ -20,6 +20,7 @@
   - 组件 `comvisual/widget/floatTool/FloatTool.vue` 新增。
 
 - `webapp-mobile` 子模块功能优化。
+  - 优化 navigation 下的 `navigationDemo.ts` 中的配置，以优化示例的展示效果。
   - 增加 VIM 组件 `VimCompreg` 的接口方法，提供更多的功能。
   - 优化 Compreg `miscellaneous.compregFallback` 的组件设置。
   - 优化 `src/views/nodes/comvisual/*/*.vue` 中占位符部分结构优化。